backend/tests/test_workflow.py
import pytest
import requests
import base64
import os
import time
from playwright.sync_api import Page, expect

BACKEND_URL = "http://localhost:5000"
FRONTEND_URL = "http://localhost:3000"
BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
IMAGES_DIR = os.path.join(BASE_DIR, "images")

if not os.path.exists(IMAGES_DIR):
    os.makedirs(IMAGES_DIR)

def get_base64_image(image_path):
    with open(image_path, "rb") as image_file:
        return base64.b64encode(image_file.read()).decode('utf-8')

# --- API TESTS ---
def test_face_verification_api():
    """
    Test the /api/verify-face endpoint with a known valid image.
    REQ: Backend must be running on port 5000.
    """
    image_path = os.path.join(BASE_DIR, "dataset", "user1.jpg")
    
    # Skip if dataset image not found
    if not os.path.exists(image_path):
        pytest.skip(f"Dataset image not found at {image_path}")

    print(f"Testing API with image: {image_path}")
    b64_img = get_base64_image(image_path)
    
    payload = {
        "image": f"data:image/jpeg;base64,{b64_img}",
        "user_id": "TEST_USER_API"
    }

    try:
        response = requests.post(f"{BACKEND_URL}/api/verify-face", json=payload)
        assert response.status_code == 200
        
        data = response.json()
        print(f"API Response: {data}")
        
        # We expect a valid match for the user in the dataset
        assert data["success"] is True
        # Depending on logic, might satisfy match or mismatch if threshold is high
        # But for 'user1.jpg' against itself (if in dataset), it should match.
        
    except requests.exceptions.ConnectionError:
        pytest.fail("Could not connect to Backend API. Is Flask running on port 5000?")

# --- UI TESTS ---
def test_ui_full_flow(page: Page):
    """
    Test the full user journey:
    1. Login (Mock)
    2. View Marketplace
    3. Verify Identity (Mock Face Scan)
    4. Buy Ticket
    5. Check Wallet
    """
    
    print("Starting UI Full Flow Test...")
    
    # 1. Login Page
    page.goto(f"{FRONTEND_URL}/login")
    page.wait_for_load_state("networkidle")
    page.screenshot(path=os.path.join(IMAGES_DIR, "01_login_page.png"))
    
    # Fill fake credentials for signup (since auth is real via supabase, 
    # we might be blocked if we don't have a real user. 
    # HOWEVER, the logs show a 'handleSignup' that just pushes to '/' on success
    # BUT we need a real session? the Page wrapper checks getSession via Supabase.
    # 
    # STRATEGY: We will try to simulate a successful login state or use valid credentials if we had them.
    # OR: We rely on the fact that this is a 'demo' environment.
    # Looking at login/page.jsx, it calls supabase.auth.signUp.
    # If we use a random email/pass, it might work if email confirmation is off.
    # Let's try to signup with random creds.
    
    email = f"testuser_{int(time.time())}@example.com"
    password = "password123"
    
    page.fill('input[type="email"]', email)
    page.fill('input[type="password"]', password)
    page.screenshot(path=os.path.join(IMAGES_DIR, "01b_login_filled.png"))
    
    # Click Sign Up
    page.click("text=Sign up")
    
    # Wait for navigation to home '/'
    # If supabase requires email verification, this might fail or stay on page.
    # We will wait and see. Increase timeout.
    try:
        page.wait_for_url(f"{FRONTEND_URL}/", timeout=10000)
    except:
        print("Signup might have failed or verify email required. Checking for error.")
        page.screenshot(path=os.path.join(IMAGES_DIR, "01c_login_error.png"))
        # If we are stuck, we can't proceed. But let's assume for 'demo' it allows access 
        # OR we can inject a session? Hard with HttpOnly cookies.
        # Let's try to click 'Sign in' with a possibly existing user?? No knowledge.
        # FOR NOW: Let's assume signup worked or we are redirected.
    
    # 2. Marketplace (Home)
    expect(page).to_have_url(f"{FRONTEND_URL}/")
    page.wait_for_selector("text=SatyaTicketing") # Navbar branding
    page.screenshot(path=os.path.join(IMAGES_DIR, "02_marketplace.png"))
    
    # Select an event. Assuming there are events rendered.
    # Look for a "Buy Ticket" or "₹" price button or card.
    # Clientside code: EventsMarketplace renders.
    # We'll click the first event card we see? Or a specific button.
    # Let's assume there is a button "Buy for ₹..." or similiar.
    # We click an element that looks clickable in the grid.
    
    # Finding an event card...
    # page.click(".event-card:first-child") # heuristic
    # Let's try to find text "Buy" or price.
    page.wait_for_timeout(2000) # wait for list hydration
    
    # Click on the first "Buy Now" or "View" button. 
    # Looking at code: The cards execute `setSelectedEvent(event)` on click?
    # Or there is a button. We'll click the first image or text block in the grid.
    # Let's try clicking a text containing "₹"
    
    # Get first element with price
    first_price = page.locator("text=₹").first
    if first_price.count() > 0:
        first_price.click()
    else:
        # Fallback: click valid-looking area
        page.mouse.click(300, 300) 
        
    page.wait_for_timeout(1000)
    page.screenshot(path=os.path.join(IMAGES_DIR, "03_event_selected.png"))

    # 3. Verification
    # Check if we need verification. Code: if !user.verified show IdentityVerification
    # We expect the "One-Time Verification" modal.
    if page.locator("text=One-Time Verification").is_visible():
        print("Verification Modal Visible")
        page.screenshot(path=os.path.join(IMAGES_DIR, "04_verification_modal.png"))
        
        # Click "Scan Biometrics to Verify"
        page.click("text=Scan Biometrics to Verify")
        
        # Wait for "Generating Biometric Hash..."
        page.wait_for_selector("text=Generating Biometric Hash...")
        page.screenshot(path=os.path.join(IMAGES_DIR, "04b_scanning.png"))
        
        # Wait for "Identity Verified" (simulated in frontend after 100% progress)
        page.wait_for_selector("text=Identity Verified", timeout=15000)
        page.screenshot(path=os.path.join(IMAGES_DIR, "04c_verified.png"))
        
        # Click "Confirm Purchase"
        page.click("text=Confirm Purchase")
    
    else:
        print("User might already be verified or modal not appeared.")
        # Try to click Confirm Purchase if it's the bottom right popup
        if page.locator("text=Confirm Purchase").is_visible():
             page.click("text=Confirm Purchase")

    # 4. Purchase Process
    # Wait for "Minting" or processing
    # page.wait_for_selector("text=Minting", timeout=5000) 
    
    # 5. Wallet View
    # After purchase, app sets view='wallet'
    # We look for "My Tickets" or wallet view indicators.
    page.wait_for_timeout(5000) # wait for async minting
    page.screenshot(path=os.path.join(IMAGES_DIR, "05_wallet_view.png"))
    
    # Verify we see a ticket
    # expect(page.locator("text=TKT-")).to_be_visible() # Heuristic for ticket ID
    
    print("Test Complete.")
