backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
deepface==0.0.75
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
pytest
pytest-playwright
tensorflow==2.15.0
tf-keras==2.15.0
